sync task.c

+++ include/task.h
@@ -1,8 +1,6 @@
 #ifndef GUARD_TASK_H
 #define GUARD_TASK_H
 
-#include "global.h"
-
 #define HEAD_SENTINEL 0xFE
 #define TAIL_SENTINEL 0xFF
 #define TASK_NONE TAIL_SENTINEL
@@ -34,7 +32,7 @@ void SwitchTaskToFollowupFunc(u8 taskId);
 bool8 FuncIsActiveTask(TaskFunc func);
 u8 FindTaskIdByFunc(TaskFunc func);
 u8 GetTaskCount(void);
-void SetWordTaskArg(u8 taskId, u8 dataElem, unsigned long value);
+void SetWordTaskArg(u8 taskId, u8 dataElem, u32 value);
 u32 GetWordTaskArg(u8 taskId, u8 dataElem);
 
 #endif // GUARD_TASK_H
